Start client async when there's state to send (#23963)
When user code deployment is enabled, the client was not starting async,
instead, it was trying to send its state to cluster and blocks until it
can do that. This is not the expected behaviour. As an async client
should not block in any way, this commit sends the client state in sync
only the client is not async.

Fixes API-1950