Fix ClientSessionManagerTest [API-2018] (#24306)
The `sessionHeartbeatsStopSent_afterClusterRestart` test failed with an
exception saying that we have made more heartbeat requests than
expected.

Looking at the test, I saw that it is easily reproducible by adding a
sleep before starting to shutdown members. In the test, we are asserting
that the number of heartbeat requests we made over the test's duration
is capped at 10, which might fail from time to time.

Besides that, the test was in the `AbstractProxySessionManagerTest`
file, which means it was executed for the member and client side proxy
session manager implementations. However, I believe it doesn't make
sense to have a member-side test, because we are trying to verify that
there are no more heartbeats sent after the cluster restart. But, we are
trying to verify that with the session manager of one of the terminated
members. I feel like we are testing the obvious - the dead member cannot
connect to the newly started cluster and send heartbeats - and that is
unnecessary. Therefore, I have moved the test to
`ClientSessionManagerTest`.

To fix the problem, I have asserted that we are doing the session
invalidation, and after that the number of heartbeats we have sent stays
the same.

closes https://github.com/hazelcast/hazelcast/issues/23310