Fix tiny/small int type handling in JdbcSqlConnector
Using `getObject` doesn't return correct type for tiny/small int due to
backwards compatibility issue, from JDBC spec:

> The JDBC 1.0 specification defined the Java object mapping for the
SMALLINT and TINYINT JDBC types to be Integer. The Java language did not
include the Byte and Short data types when the JDBC 1.0 specification was
finalized. The mapping of SMALLINT and TINYINT to Integer is maintained to
preserve backwards compatibility.