Support DML predicates in JdbcSqlConnector [HZ-2084] (#24521)
Adds support for predicates in DML (update, delete) for the
jdbcSqlConnector.

If an expression is supported is decided by the `SupportsRexVisitor`,
which traverses the RexNode to see if all nested expressions are
supported.

When the predicate can be pushed down and there is no input, the
WriteJdbcP would execute nothing. A fake input source vertex is provided
and it produces a single fake row.

The `UpdateProcessorSupplier` and `DeleteProcessorSupplier` were unified
into `DmlProcessorSupplier` as they were identical. The dynamic
parameters and input refs are collected separately.

New test coverage:
- PredicatePushDownJdbcSqlConnectorTest which tests that expressions
from
https://docs.hazelcast.com/hazelcast/5.2/sql/functions-and-operators#hide-nav
can be pushed down. It runs the actual query and also explain to verify
the predicate was pushed down. Runs against H2 (in PR builder) and
Postgres + MySQL (nightly).
-
UpdatePredicatePushDownJdbcSqlConnectorTest/DeletePredicatePushDownJdbcSqlConnectorTest
which test some edge cases for update and delete

Co-authored-by: Krzysztof Jamróz <krzysztof.jamroz@hazelcast.com>