Decrease local parallelism of processors in JdbcSqlConnector [HZ-1637] (#24534)
The call to `Map#clear` with enabled `MapStore` results in at least 1
call to `MapStore#deleteAll` for each partition. The ultimate fix would
be to add `MapStore#deleteAll` method, but the benefit of such API in
MapStore is unclear.

However, it made the following issue visible:

When not explicitly set via forceTotalParallelismOne or
setLocalParallelism the local parallelism default value is larger than
1. This results in 2 or more processors being created for each
insert/update/delete/sink into. Each processor requests a connection
from a pool. It is possible that 2 concurrently running queries borrow 1
connection each, exhausting the pool and resulting in a deadlock, which
ends after a 30 s timeout (default) from the connection pool.

It is arguably a user error that the connection pool size is too low,
but with localParallelism>1 it is really easy to hit this issue (as seen
in the `Map#clear`)

Also, it is theoretically possible to hit this issue across multiple
members (the processor on one member successfully obtains a connection
from the pool, but waits on another member and vice versa), but I didn't
manage to reproduce this, so I didn't focus on fixing that issue.

In the future, we can make this behaviour smarter for insert/sink into
or update/delete which have input and could make use of higher local
parallelism.

Fixes #22567