Fix TINY/SMALL INT type handling in JdbcSqlConnector (#24535)
Using `getObject` doesn't return the correct type for tiny/small int due
to backwards compatibility issue, from JDBC spec:

> The JDBC 1.0 specification defined the Java object mapping for the
SMALLINT and TINYINT JDBC types to be Integer. The Java language did not
include the Byte and Short data types when the JDBC 1.0 specification
was finalized. The mapping of SMALLINT and TINYINT to Integer is
maintained to preserve backwards compatibility.