Add lifecycle listener before client starts in ConfiguredBehaviorTest
In the `testAsyncStartTrueShouldNotBlock_whenThereIsClientStateToSend`
we add a lifecycle listener after the client starts, and wait
until the client moves into the connected state.

However, the client might move to the connected state even before
we have added that listener, as the listener is added after
the client starts.

I have moved the listener registration to the configuration so that
we would receive the connected event for sure.