Verify JoinOperation can shutdown [HZ-2429] (#24577)
`JoinOperation`s are not authenticated. So if they do a node shutdown,
a 3rd party can easily shutdown the node. To prevent this, we do not
allow shutting down the node if it ever joined a cluster.

`WanOperation`s are also susceptible to this but I couldn't find any
shutdown there.

ShutdownNodeOp also does a shutdown and JoinOperation. But I can't
disallow shutting down there because that operation needs to shutdown
the node even if the node is joined. It requires a passive cluster
state, so maybe it's a lesser problem than the other operations.

Fixes https://hazelcast.atlassian.net/browse/HZ-2429