[TPC] Drop ifSupported methods in AsyncSocketOptions (#24615)
The new AsyncSocketOptions API offers three methods: isSupported, get
and set.
isSupported allows to check whether the option is supported. Set allows
to
assign a value to an option and get allows to retrieve the value that
was
assigned earlier. If the option is not supported, the method call causes
no side effects and no exception is thrown. Furthermore, the set
operation
returns a boolean to acknowledge whether the option is supported and the
value
was assigned.