Remove CA_CERTIFICATE declaration in KubernetesClientTest
With the old RestClient implementation, the connection formed in this test is never `HttpsURLConnection`, so the certificate was never used. In the new RestClient, the SSL context is created before the connection, so passing an invalid certificate breaks functionality.

Since this test did not utilise the caCertificate value before, there should be no impact from removing it.