Replace Mockito spy usage in QueryIndexingTest [HZ-2491] (#24778)
We have recently been encountering a variety of edge-case issues with
Mockito related failures in some of our tests - these tests are
generally non-deterministic, but surface during long-running test jobs
on Jenkins. There has been extensive investigation of these issues by
multiple team members, but the issues always point towards an issue
within Mockito itself (although exact causes have not been identified).

The QueryIndexingTest encounters the same failures consistently while
running on a Sonar code coverage Jenkins job. The failures are always
related to instantiating a _serialized spy_ instance, with various
debugging sessions showing that non-serialized spy instances are created
successfully. As with other similar Mockito failures, this failure is
reproduced consistently in long-running tests, but not in isolation.

Due to the functionality of this being simple in this test, we have
decided to avoid spending more time investigating the specifics of why
Mockito is struggling here and simply replace the problematic Mockito
functionality with our own implementation. In this test that is a simple
wrapper for the `Employee` class which uses a `Map` keyed by method
names to an `AtomicInteger` which counts the invocations of each method.

Fixes https://github.com/hazelcast/hazelcast/issues/23645