Introduce latch and more logging to RestClientTest TLS 1.3 support tests
There are two tests within the `RestClientTest` class which have been
shown to non-deterministically fail occasionally during Jenkins job;
those tests are `tls13SupportDefaultCacert` and `tls13SupportCustomCacert`.

During my debugging I was able to reproduce the failures only as part of
a long-running test suite, never as an isolated test. I was only able to
reproduce the issue once with additional logging, which came in the form
of a single `"TLS 1.3 found? <boolean>"` logger call after the cipher had
been checked - the logging output showed this being called twice with `true`
values both times, which should indicate the test was successful. However,
the test had still failed due to the assertion check at the end of the test.

Any attempts to debug further resulted in being unable to reproduce the
failure, even during long-running tests, leading me to believe this is the
result of a niche race condition, most likely around the socket closure,
as we know at least this stage is reached before there is divergence.

This commit primarily introduces a `CountDownLatch` for each of the
`Tls13CipherCheckingServer` instances, which is counted down when the thread
has completed, and is awaited before the test asserts whether the cipher was
found. This allows us to guarantee that the cipher thread has completed (or
at the very least will show a 5s timeout in logs, indicating further issues)
before we check to see if the TLS 1.3 cipher was found. This should stop us
from encountering this niche race condition in our tests.

This commit also contains some additional logging verbosity as well as some
very minor code cleanup related to the TLS 1.3 cipher check tests.