Improvements to RestClientTest TLS 1.3 support tests [HZ-2541] (#24794)
Fixes https://github.com/hazelcast/hazelcast/issues/24676

There are two tests within the `RestClientTest` class which have been
shown to non-deterministically fail occasionally during Jenkins job;
those tests are `tls13SupportDefaultCacert` and
`tls13SupportCustomCacert`.

This PR improves error handling, synchronization and logging in the
tests.

Notes:
> During my debugging I was able to reproduce the failures only as part
of a long-running test suite, never as an isolated test. I was only able
to reproduce the issue once with additional logging, which came in the
form of a single `"TLS 1.3 found? <boolean>"` logger call after the
cipher had been checked - the logging output showed this being called
twice with `true` values both times, which should indicate the test was
successful. However, the test had still failed due to the assertion
check at the end of the test.
>
> Any attempts to debug further resulted in being unable to reproduce
the failure, even during long-running tests, leading me to believe this
is the result of a niche race condition, most likely around the socket
closure, as we know at least this stage is reached before there is
divergence.

---------

Co-authored-by: Josef Cacek <josef.cacek@gmail.com>