Replace Mockito invocation counter in MetricsCompressorTest [HZ-2538] (#24773)
We have recently been encountering a variety of edge-case issues with
Mockito related failures in some of our tests - these tests are
generally non-deterministic, but surface during long-running test jobs
on Jenkins. There has been extensive investigation of these issues by
multiple team members, but the issues always point towards an issue
within Mockito itself (although exact causes have not been identified).

The MetricsCompressorTest encounters the same failures for the same set
of tests, across multiple JVM vendors and versions. The failure is
always related to the `verify(supplierSpy, times(2)).get();` check, with
`times(2)` being `only()` in some tests (although still failing).

Due to the functionality of this being very simple in this test, we have
decided to avoid spending more time investigating the specifics of why
Mockito is struggling here and simply replace the problematic Mockito
functionality with our own implementation. In this test that is a simple
wrapper for the `DefaultMetricDescriptorSupplier` which uses an
`AtomicInteger` to count the invocations on the `#get()` method.

Fixes https://github.com/hazelcast/hazelcast/issues/23504