Update persistence metric naming convention to use discriminators
Currently, persistence related metrics are recorded with the prefix
`persistence.<store_name>.<metric>` - the inclusion of `<store_name>`
here is not good practice and also not consistent with other metrics
(such as operation related metrics). The `store_name` component is not
static and should be included as a tag/label as opposed to being
integrated into the metric name itself. This allows the metric to play
nicely with Prometheus and adheres to [standard naming practices](https://prometheus.io/docs/practices/naming/).

This commit resolves this by changing the `store_name` aspect to be a
metric discriminator with the tag `storeName`. The `store_name` itself
is compromised of the instance name and chunk sequence ID - I considered
splitting this up further into separate tags, but I decided against it
as the chosen approach seems to follow convention used in other metrics
such as the `thread` discriminator in operation metrics.

Fixes https://hazelcast.atlassian.net/browse/HZ-1299