Update metrics for CPSession and Lock in CP Subsystem
Currently there is some inconsistency in the way metrics are handled
for `hz_cp_session_*` and `hz_cp_lock_ownerSessionId` - the CPSession
follows the CP subsystem `id` naming convention of using its own ID
(session ID) combined with its group name (using the format `id@group`).
However this makes it difficult to join the metrics of CPSessions with
those of owned Lock metrics, as there is no common tag to join with.

This commit aims to resolve this issue by including 2 new tags in the
metrics for both CP Sessions and Locks: `sessionId`, which is the
session's ID for Sessions and the owning session's ID for Locks; and
`qualifiedSessionId` which is the same as `sessionId`, with the group
qualifier applied in the standard format (`sessionId@groupName`). These
new tags allow for Session and owned Lock metrics to be joined easily.

I decided to have Lock metrics continue to provide `owner.sessionId()`
as the value of the `hz_cp_lock_ownerSessionId` metric; although we now
include this data in tags, continuing to provide this metric this way
provides backwards compatibility, and there is no significant meaningful
data to provide in its place related to the `owner` - alternatively we
simply drop this metric as it no longer provides much value for us.

Fixes https://hazelcast.atlassian.net/browse/HZ-2081