Drop Lite member check in LocalMapStatsProvider, use MapProxyImpl stats
Based on Ahmet's feedback from my PR - testing suggests that after a Lite
member invokes a `PutOperation`, it does get a `MapContainer` object,
meaning that cleanup of the `LocalMapStatsImpl` is already handled in the
existing `MapServiceContextImpl#destroyMap` function. I made an addition
to my regression test which tests for this and confirms this destruction.