Simplify stats fetch in LocalMapStatsProvider and cleanup on #destroyMap
The `MapProxySupport#getLocalMapStats` method is creative, so if there
was no `LocalMapStatsImpl` created prior, it would be created - also
under the hood it calls back to the `LocalMapStatsProvider` to check
its `statsMap` field anyway, so we can cut out the middle man and
check the `statsMap` field directly. If no existing stats exist in the
`statsMap` map, then an empty `LocalMapStatsImpl` is used instead -
this maintains any prior functionality related to "empty stats".

My testing also discovered that our current `MapServiceContext#destroyMap`
method does not account for Lite members that never access a Map (which
means they never get a `MapContainer`), but do get a `LocalMapStatsImpl`
created when their map proxy is created. As they have no `MapContainer`,
the `MapServiceContext#destroyMap` method does not reach the calls where
the stats are cleaned up - this commit resolves that by calling the
removal function even if there is no `MapContainer` present.