Provide LocalMapStats data for invocations from Lite members [HZ-2072] (#24871)
**Context:** Metrics for IMaps are stored in `LocalMapStats` objects,
which are provided by `LocalMapStatsProvider`. Certain MapOperation
metrics (get, set, put, & remove) are logged when
`MapProxySupport#invokeOperation` is called, which uses
`MapOperationStatsUpdater.incrementOperationStats`. This happens for
operations not executed on the local member as well, which means that a
Lite member invoking `IMap#put` will cause a `PutOperation` being
executed on another (non-Lite) member, but it will increment stats on
its own member.

**Problem:** Due to Lite members not owning any data themselves, they do
not report any metrics for MapOperations. This means that metrics on
these operations are lost with the current metric collection mechanics.

**Solution:** Tweaking
`LocalMapStatsProvider#addStatsOfNoDataIncludedMaps` to provide
the local `statsMap` value of `LocalMapStatsImpl` when there is no data
currently included - this `LocalMapStatsImpl` contains the invocation
counts from Lite members (which is already correctly updated). This
is only done if an empty `LocalMapStats` object would've been passed
otherwise, and the `statsMap` value is not null.

While testing I also identified a scenario where a Lite member can have
its own `LocalMapStatsImpl` for a map which it does not currently hold
a `MapContainer` for - this means the stats object is not cleaned up. I
made changes to `MapServiceContextImpl#destroyMap` to account for
this and clean up these objects on Lite members.

I have also added a very simple regression test which ensures Lite
members produce metrics after `IMap#put` is called.

**Note:** It may be worth reworking our map statistic capturing
mechanic, shifting the responsibility to the executing member rather
than the invoker. I chose my approach with the mindset of changing the
least amount of existing functionality/behaviour, while also stopping
these metrics being lost entirely as they currently are.

Closes https://hazelcast.atlassian.net/browse/HZ-2072